2022 Day 16 & 19: Optimizations

+++ 2022/day16/src/lib.rs
@@ -9,14 +9,14 @@ mod state;
 
 aoc_tools::aoc_sol!(day16 2022: part1, part2);
 
-pub fn part1(input: &str) -> u32 {
+pub fn part1(input: &str) -> u16 {
     let network = parse_input(input);
     let mut solver = Solver::part1(&network);
     while solver.advance(&network) {}
     solver.best_lower_bound
 }
 
-pub fn part2(input: &str) -> u32 {
+pub fn part2(input: &str) -> u16 {
     let network = parse_input(input);
     let mut solver = Solver::part2(&network);
     while solver.advance(&network) {}
